Fix unix socket issue
Previously, we've been failed to communicate through unix socket.
So, we choose to port as alternation.
But, by using volume properly, unix socket communication has become possible.

`./flask/socket` is a host path and `/app/socket` is container path. So, for sharing socket file, we need to assign same path in host path.